IDEA-157062 AIOOBE at StandardGlyphVector.getLogicalBounds
This adjusts the fix for IDEA-156971, so that original font's style is used to create a native font object. It turns out, changing style does have an impact on char-to-glyph mapping in native code.
This can influence also fake-italic fonts created by CFontManager, but it's only critical for our case, where we must use exactly the same font which is returned by font fallback code.